pyproject.toml
[tool.poetry]
name = "yamyam-lab"
version = "0.1.0"
description = ""
authors = ["skip"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
scikit-learn = "1.5.2"
torch = "2.5.1"
pandas = "2.2.3"
hydra-core = ">=1.3.2"
torch-geometric = ">=2.6.1"
lightgbm = ">=4.5.0"
xgboost = ">=2.1.2"
ipykernel = ">=6.29.5"
black = ">=24.10.0"
flake8 = ">=7.1.1"
isort = ">=5.13.2"
prettytable = ">=3.12.0"
seaborn = ">=0.13.2"
matplotlib = "3.9.2"
implicit = "0.7.2"
node2vec = ">=0.5.0"
flake8-pyproject = ">=1.2.3"
pyg_lib = { url = "https://data.pyg.org/whl/torch-2.5.0%2Bcpu/pyg_lib-0.4.0%2Bpt25-cp311-cp311-macosx_14_0_universal2.whl" }
torch_scatter = { url = "https://data.pyg.org/whl/torch-2.5.0%2Bcpu/torch_scatter-2.1.2-cp311-cp311-macosx_10_9_universal2.whl" }
torch_sparse = { url = "https://data.pyg.org/whl/torch-2.5.0%2Bcpu/torch_sparse-0.6.18-cp311-cp311-macosx_11_0_universal2.whl" }
torch-cluster = { url = "https://data.pyg.org/whl/torch-2.5.0%2Bcpu/torch_cluster-1.6.3-cp311-cp311-macosx_10_9_universal2.whl" }
torch_spline_conv = { url = "https://data.pyg.org/whl/torch-2.5.0%2Bcpu/torch_spline_conv-1.2.2-cp311-cp311-macosx_10_9_universal2.whl" }


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 119

[tool.black]
line-length = 119
target-version = ['py36', 'py37', 'py38']
exclude = '''
    \.git
  | \.mypy_cache
  | \.tox
  | venv
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | ^.*\b(migrations)\b.*$
'''

[tool.flake8]
max-line-length = 119
ignore = "E203, W503, E501"
exclude =  '''
    .git,
    .gitignore,
    */migrations/*,
    __pycache__,
'''